backend/requirements.txt
anyio>=3.5.0
asgiref>=3.5.0
backcall>=0.2.0
beautifulsoup4>=4.10.0
bs4>=0.0.1
certifi>=2021.10.8
charset-normalizer>=2.0.12
click>=8.1.2
colorama>=0.4.3
decorator>=4.4.2
fastapi>=0.75.1
h11>=0.13.0
idna>=3.3
ipython>=7.18.1
ipython-genutils>=0.2.0
jedi>=0.17.2
jupyter-core>=4.6.3
parso>=0.7.1
pickleshare>=0.7.5
prompt-toolkit>=3.0.7
pydantic>=1.9.0
Pygments>=2.7.1
python-dateutil>=2.8.1
requests>=2.27.1
six>=1.15.0
sniffio>=1.2.0
soupsieve>=2.3.1
starlette>=0.17.1
tornado>=6.0.4
traitlets>=5.0.4
typing_extensions>=4.1.1
urllib3>=1.26.8
uvicorn>=0.17.6
wcwidth>=0.2.5
alembic>=1.8.0
pymysql>=1.0.2
email-validator>=1.0.5
passlib>=1.7.4
bcrypt>=3.2.2
python-jose[cryptography]>=3.3.0
pydantic[dotenv]>=1.9.0
python-jose[cryptography]>=3.3.0
pydantic[dotenv]>=1.9.0
fastapi-pagination>=0.10.0